+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Next
 
+- **[Breaking change]** Rename `typed-enum` to `white-list`
 - **[Breaking change]** Split serialization functions by format to prepare tree-shaking (#39)
 
 # 0.5.0 (2018-01-02)